# Add prepared statement caching for SQLite