--- a/diesel/src/expression/functions/aggregate_ordering.rs
+++ b/diesel/src/expression/functions/aggregate_ordering.rs
@@ -40,6 +40,10 @@ macro_rules! ord_function {
                 try!(self.target.collect_binds(out));
                 Ok(())
             }
+
+            fn is_safe_to_cache_prepared(&self) -> bool {
+                self.target.is_safe_to_cache_prepared()
+            }
         }
 
         impl<T: Expression, QS> SelectableExpression<QS> for $type_name<T> {