--- a/diesel/src/query_builder/distinct_clause.rs
+++ b/diesel/src/query_builder/distinct_clause.rs
@@ -15,6 +15,10 @@ impl<DB: Backend> QueryFragment<DB> for NoDistinctClause {
     fn collect_binds(&self, _out: &mut DB::BindCollector) -> QueryResult<()> {
         Ok(())
     }
+
+    fn is_safe_to_cache_prepared(&self) -> bool {
+        true
+    }
 }
 
 impl<DB: Backend> QueryFragment<DB> for DistinctClause {
@@ -25,5 +29,9 @@ impl<DB: Backend> QueryFragment<DB> for DistinctClause {
 
     fn collect_binds(&self, _out: &mut DB::BindCollector) -> QueryResult<()> {
         Ok(())
+    }
+
+    fn is_safe_to_cache_prepared(&self) -> bool {
+        true
     }
 }