--- a/diesel/src/query_builder/update_statement/mod.rs
+++ b/diesel/src/query_builder/update_statement/mod.rs
@@ -65,6 +65,10 @@ impl<T, U, DB> QueryFragment<DB> for UpdateStatement<T, U> where
         }
         Ok(())
     }
+
+    fn is_safe_to_cache_prepared(&self) -> bool {
+        false
+    }
 }
 
 impl<T, U> AsQuery for UpdateStatement<T, U> where
@@ -152,5 +156,9 @@ impl<T, U, DB> QueryFragment<DB> for UpdateQuery<T, U> where
         try!(self.statement.collect_binds(out));
         try!(self.returning.collect_binds(out));
         Ok(())
+    }
+
+    fn is_safe_to_cache_prepared(&self) -> bool {
+        false
     }
 }