This is the first pass at adding prepared statement caching to Diesel.
The query builder will still run, as we are using the SQL string as the
key. Eventually this will be replaced with a type identifier, but we're
not there yet.

Most of the groundwork is laid to apply this to PG as well, but to start
it only applies to SQLite. The cache is placed at the top of the struct
instead of the end to ensure that the statements are dropped before the
raw connection. I'm still trying to determine if the order that fields
are dropped is something we can rely on. We may need to come up with an
alternate implementation for that part of this.

The prepared statements are stored in a hash map, and not some sort of
LRU cache. This means that the memory growth here could be unbounded if
left unchecked. In practice though, there is an upper bound as long as
we exclude certain types of queries from the cache. In particular, we
need to exclude queries that can generate different SQL without the type
changing. Therefore, we apply the following rules:

- ASTs containing the `Many` node are not cached (`WHERE foo IN (?, ?,
  ?)` type queries)
- ASTs containing a `SqlLiteral` node are not cached
- Insert statements are not cached
- Update statements are not cached

There's an argument for allowing `SqlLiteral` nodes to be cached, but I
am erring on the side of caution here. We can't tell if you're
dynamically generating the SQL in that node or not, and allowing them to
be cached would require exposing controls to care about internals that I
want to keep internal for the time being.

This results in a healthy performance improvement across the board.

Before
------

```
test bench_selecting_0_rows_with_medium_complex_query         ... bench:      43,125 ns/iter (+/- 12,545)
test bench_selecting_0_rows_with_trivial_query                ... bench:       5,755 ns/iter (+/- 2,063)
test bench_selecting_10k_rows_with_medium_complex_query       ... bench:  22,090,236 ns/iter (+/- 7,078,394)
test bench_selecting_10k_rows_with_medium_complex_query_boxed ... bench:  20,322,178 ns/iter (+/- 5,829,031)
test bench_selecting_10k_rows_with_trivial_query              ... bench:   2,980,795 ns/iter (+/- 1,046,410)
test bench_selecting_10k_rows_with_trivial_query_boxed        ... bench:   2,894,526 ns/iter (+/- 829,044)
```

After
-----

```
test bench_selecting_0_rows_with_medium_complex_query         ... bench:      16,414 ns/iter (+/- 4,357)
test bench_selecting_0_rows_with_trivial_query                ... bench:       1,187 ns/iter (+/- 296)
test bench_selecting_10k_rows_with_medium_complex_query       ... bench:  19,490,094 ns/iter (+/- 4,403,771)
test bench_selecting_10k_rows_with_medium_complex_query_boxed ... bench:  18,980,924 ns/iter (+/- 3,537,891)
test bench_selecting_10k_rows_with_trivial_query              ... bench:   2,873,268 ns/iter (+/- 1,174,266)
test bench_selecting_10k_rows_with_trivial_query_boxed        ... bench:   2,891,690 ns/iter (+/- 1,012,551)
```

We can probably structure this a little bit more "unsafely" to eliminate
the overhead of `Rc` and `RefCell`. A statement will never have more
than one reference at a time (and this is enforced by `RefCell`, since
basically every method on `Statement` requires `&mut self`). We could in
theory just "copy" the `Statement` struct instead. The code is
encapsulated to allow this, but the overhead of `Rc<RefCell<>>` isn't
high enough to justify it.