@samphippen 2016-04-17
Your tests here seem legit, but you've got like a billion fiddly little boolean cases, are there any important ones that are missing?