@sgrif 2016-04-17
I don't think so. The only important cases are the ones that return the literal `false`. The tests there were more to drive implementation than enforce anything specific. Perhaps I need to change the implementation to make this more clear, but all the "fiddly" cases are basically saying "I'm safe to cache if all of my components are safe to cache"